Introduced a bind on non local variable
With this change, you can adapt a variable to allow
all the apps consuming keepalived to bind on an address
that isn't defined on the system yet (for example if
you don't have the vip because you're backup,  you can
still start your app that binds on the vip).

Please check readme file.