Introduced Ubuntu Cloud Archive installation method
Instead of using a ppa, some deployers may want to use ubuntu cloud
archive to install keepalived.

This commit introduces two variables to change the default behavior (of
installing with ppa): keepalived_uca_enable and uca_apt_repo_url

The first variable is defaulted to false, and if set to true, keepalived
will be installed with ubuntu cloud archive.
The second variable is the url to the uca repository, which can be
overriden if you want to define your own mirror of the uca repository
(Defaults to: "http://ubuntu-cloud.archive.canonical.com/ubuntu")

Signed-off-by: Jean-Philippe Evrard <jean-philippe@evrard.me>