Add SELinux module for keepalived's ping cmd
The `ping` command used by keepalived causes SELinux AVC denials on
CentOS 7. This causes keepalived to think that all of the systems
are unstable and it can't find a suitable host to run services.

This patch adds a policy that allows keepalived to function without
causing AVCs.

For more details:
  https://bugs.launchpad.net/openstack-ansible/+bug/1660426