Adding timeout attribute for check script (#41)
Merge branch 'master' of https://github.com/BjoernT/ansible-keepalived.

This is modified to solve merge conflicts and still keep the
features asked. It generally is less intrusive for updates, as this won't change
the config file by default.

This brings a new feature, to add timeout on vrrp check scripts.