Allow list of SELinux compile rules
Keepalived role doesn't need HAProxy pid SELinux permissions.
Therefore the keepalived_haproxy_pid_file was removed from
the list of rules to compile.

However, flexibility could be advisable, and the list of rules
to compile should not be static, but instead be an overridable
list. By default, the list will compile the ping rule, as it
was done before, for retro-compatibility.

But a deployer can now provide path to files to compile.