Allow list of SELinux compile rules
This patch breaks the SELinux compilation part of the role
into a reusable set of tasks.

The list of rules to compile is not static, but is instead
an overridable list. By default, the list will compile the
ping rule, as it was done before, for retro-compatibility.

A deployer can now provide path to files to compile.

OpenStack-Ansible users can use this patch to
allow keepalived to cat haproxy pid.