Fix galaxy metadata
Without this patch, galaxy recognizes the role as
"ansible-keepalived" instead of "keepalived", and the author as
my full name. This behaviour changed in Galaxy 3, where the
automatic wiring of `ansible-role-x` or `ansible-x` to `x`
disappeared.

This is a problem, as galaxy is now unable to recognize the role,
and guesstimates the galaxy namespace to be "ansible-keepalived",
which is invalid, as it contains a dash.

This fixes it by ensuring the namespace is "evrardjp" (my github
account), and the role is named `keepalived`, which should re-wire
the role to its proper location.