Split ansible-2.10 requirements from dev
The ansible package ansible==2.10 is NOT having the proper
constraint to matching ansible-base/ansible-core. So it means
that pip happily installs ansible==2.10 then installs
ansible-core 2.13. This means the code actually running
in ansible-2.10 testing is ansible-2.13.

This is a problem, as the testing for ansible-2.10 is not
actually testing 2.10.

This fixes it by splitting the two test environments, having two
different requirements files. We do not lock or constrain
(yet!) the dependencies, as this would mean extra effort
in the tooling.

In the meantime, contributors should ensure that the right
version of ansible-core/base is installed.