Update include_vars to recent practices
Without this patch, the include_vars is relying on the full hostvars
which can slow down the role execution.

Instead, load the right (recommended) variables.

At the same time, the use a first_found lookup instead of a
with_first_found loop allows the 'paths' parameter to be used.

This ensures that only vars from the role are included, and not vars
from a parent calling role. This can happen when a parent role has
a higher priority vars file available for inclusion than the role
it calls.