Remove dependencies updates for old ansible
The intention of the ansible 2.10 requirements is to test
if the role is still working with old ansible.
It doesn't make sense to update this requirements file.

Yet, dependabot wants to always update it, and there is
no way to tell dependabot to ignore this requirement file
or only select the other file as the important one to update.
(unless I start to craft folders for each of those, which
is not necessary here). See also this issue in dependabot [1].

Using the .text extension has worked for mozilla [2], so I will
follow the same approach.

[1]: https://github.com/dependabot/dependabot-core/issues/1657
[2]: https://github.com/mozilla/fx-private-relay/pull/2641