Ensure locale is set to en_US for CI
Without this, it is possible for a user to use its own locale,
which could lead to ansible failure if the locale is not well
generated/accepted by ansible/python (locale.getlocale()).

To ensure there is no empty or incorrect locale in CI, force it
to en_US, which the locale used in the communication for this
project.