Implement track_process (#278)
Without this patch, a user is forced to use track_script with pgreps or kills to track process

This is not a problem in itself, but is error prone.

This patch fixes it by exposing in the role the track_process feature of keepalived.

The initial implementation was validating some of the track_process inputs directly in the role,
under the form of `param_match in ['initial', 'partial']`. 
Yet, this provides a different result (the incorrect data is silently not templated,
and no error is thrown) as a deployer would expect.

Instead, we rely on the fact that keepalived will throw a warning, and the user can
check it if necessary.